refactor: merge vision parser families back together
